# swagger_hub_api_mock.yml
openapi: 3.0.0
info:
  description: |
    Gelado, sua vida mais refrescante.
  version: "1.0.0"
  title: Gelado store
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: leonardorodrigues_f@hotmail.com
    name: Leonardo Rodrigues
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Leorfk/gelado/1.0.0
...
paths:
...
  /usuario:
    get:
      tags:
        - usuario
      summary: retorna todos os usuarios
      description: retorna uma lista de usuários
      operationId: getAllUsuario
      responses:
        '200':
          description: Operação bem sucedida
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/usuario'
        '400':
          description: Id inválido
        '404':
          description: Usuário não encontrado
      security:
        - api_key: []
    post:
      tags:
        - usuario
      summary: Cria um novo usuario
      description: Realiza o cadastro de um usuario
      operationId: createUsuario
      responses:
        default:
          description: Operação bem sucedida
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/usuario'
        description: Usuário criado
        required: true
      security:
      - api_key: []
...
components:
  schemas:
...
    usuario:
      type: object
      properties:
        idUsuario:
          type: integer
          format: int64
        email:
          type: string
        senha:
          type: string
        user_role:
          $ref: '#/components/schemas/user-role'
#  objeto role
    user-role:
      type: object
      properties:
        roleId:
          type: integer
          format: int64
        textoRole:
          type: string
...
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header